+++ utils/ablation.py
@@ -1,56 +1,21 @@
+
 import gc
-import json
-import shutil
 import torch
+import shutil
+import json
 from pathlib import Path
 from typing import Dict, List, Optional, Tuple, Any, Union
 from tqdm import tqdm
 from safetensors.torch import load_file, save_file
-from transformers.utils import cached_file
 
 from utils.config import ModelConfig
-from utils.modifier import modify_tensor_norm_preserved, modify_tensor_simple
-from utils.sparsify import percentile_sparsify, sparsify_tensor
-from utils.vector_utils import remove_orthogonal_projection
-
-def resolve_model_paths(model_id: str) -> Tuple[Optional[str], Union[Path, str], Dict[str, str], List[str]]:
-    """
-    Resolves model paths, handling both sharded and single-file models.
-    Returns: index_path, model_dir, weight_map, shards
-    """
-    index_path = cached_file(model_id, "model.safetensors.index.json")
-    model_dir = None
-    weight_map = {}
-    
-    if not index_path:
-        # Fallback for non-sharded models (single file)
-        single_file = cached_file(model_id, "model.safetensors")
-        if single_file:
-            print("Model is not sharded (single file). Treating as single shard.")
-            weight_map = {"model.safetensors": "model.safetensors"}
-            model_dir = Path(single_file).parent
-        else:
-             # Try checking if it's a local directory that doesn't use cache
-            local_path = Path(model_id)
-            if (local_path / "model.safetensors.index.json").exists():
-                 index_path = str(local_path / "model.safetensors.index.json")
-                 model_dir = local_path
-            elif (local_path / "model.safetensors").exists():
-                 print("Found local single safetensors file.")
-                 weight_map = {"model.safetensors": "model.safetensors"}
-                 model_dir = local_path
-                 index_path = None
-            else:
-                 raise ValueError("Could not find model.safetensors.index.json or model.safetensors")
-    
-    if index_path:
-        with open(index_path) as f:
-            index = json.load(f)
-        weight_map = index["weight_map"]
-        model_dir = Path(index_path).parent
-
-    shards = sorted(list(set(weight_map.values())))
-    return index_path, model_dir, weight_map, shards
+from utils.io import resolve_model_paths, copy_model_artifacts
+from utils.math_utils import (
+    modify_tensor_norm_preserved, 
+    modify_tensor_simple, 
+    sparsify_tensor,
+    remove_orthogonal_projection
+)
 
 def get_layer_ablation_config(
     config: ModelConfig,
@@ -151,42 +116,6 @@ def modify_shard_weights(
             modified = True
             
     return modified
-
-def copy_model_artifacts(config: ModelConfig, output_path: Path, index_path: Optional[str]):
-    """
-    Copies configuration and index files to the output directory.
-    """
-    print("\n" + "="*60)
-    print("PHASE 3: Finalizing")
-    print("="*60)
-    print("Copying configuration files...")
-    
-    if index_path:
-        shutil.copy(index_path, output_path / "model.safetensors.index.json")
-    
-    config_files = [
-        "config.json", "generation_config.json", 
-        "tokenizer_config.json", "tokenizer.json", "special_tokens_map.json",
-        "vocab.json", "merges.txt", "added_tokens.json", "preprocessor_config.json",
-        "chat_template.json"
-    ]
-    
-    for filename in config_files:
-        try:
-            # Wrap in try-except to simple skip missing files
-            path = cached_file(config.model_id, filename)
-            if path and Path(path).exists():
-                shutil.copy(path, output_path / filename)
-            else:
-                # Check local if cached_file returns None or fail
-                # But cached_file handles local paths too.
-                # If model_id is local path, check directly
-                if (Path(config.model_id) / filename).exists():
-                     shutil.copy(Path(config.model_id) / filename, output_path / filename)
-        except Exception as e:
-            # Just log and skip
-            print(f"  Skipping {filename}")
-            pass
 
 def run_sharded_ablation(
     config: ModelConfig,
